chore: Move Proxy & ProxyV2 to separate file
