+++ lib/clerk/rack_middleware_v2.rb
@@ -3,92 +3,6 @@ require_relative "authenticate_context"
 require_relative "authenticate_request"
 
 module Clerk
-  class ProxyV2
-    CACHE_TTL = 60 # seconds
-
-    attr_reader :session_claims, :session_token
-
-    def initialize(session_claims: nil, session_token: nil)
-      @session_claims = session_claims
-      @session_token = session_token
-      @session = nil
-    end
-
-    def session
-      return nil if @session_claims.nil?
-
-      @session ||= verify_session
-    end
-
-    def verify_session
-      return nil if @session_claims.nil?
-
-      sdk.sessions.verify_token(@session_claims["sid"], @session_token)
-    end
-
-    def user
-      return nil if user_id.nil?
-
-      @user ||= fetch_user(user_id)
-    end
-
-    def user_id
-      return nil if @session_claims.nil?
-
-      @session_claims["sub"]
-    end
-
-    def org
-      return nil if org_id.nil?
-
-      @org ||= fetch_org(org_id)
-    end
-
-    def org_id
-      return nil if user_id.nil?
-
-      @session_claims["org_id"]
-    end
-
-    def org_role
-      return nil if @session_claims.nil?
-
-      @session_claims["org_role"]
-    end
-
-    def org_permissions
-      return nil if @session_claims.nil?
-
-      @session_claims["org_permissions"]
-    end
-
-    private
-
-    def fetch_user(user_id)
-      cached_fetch("clerk_user:#{user_id}") do
-        sdk.users.find(user_id)
-      end
-    end
-
-    def fetch_org(org_id)
-      cached_fetch("clerk_org:#{org_id}") do
-        sdk.organizations.find(org_id)
-      end
-    end
-
-    def cached_fetch(key, &block)
-      if store = Clerk.configuration.middleware_cache_store
-        store.fetch(key, expires_in: CACHE_TTL, &block)
-      else
-        yield
-      end
-    end
-
-    def sdk
-      @sdk ||= Clerk::SDK.new
-    end
-  end
-
   class RackMiddlewareV2
     def initialize(app)
       @app = app
